+++ core/changelog.md
@@ -1,4 +1,6 @@
 <!-- The pattern we follow here is to keep the changelog for the latest version -->
 <!-- Old changelogs are automatically attached to the GitHub releases -->
 
-- fix: responses tool message output struct overlapping fields fixed
+- fix: openai specific parameters filtered for openai compatibile providers
+- fix: error response unmarshalling for gemini provider
+- BREAKING FIX: json_schema field correctly renamed to schema; ResponsesTextConfigFormatJSONSchema restructured